+++ NEWS.md
@@ -12,6 +12,9 @@
 * Added native triplet support in note info using the `t`-prefix, e.g., `4 4] t8 t8- t8^ 4`. Support extends to `music` objects and now also to `phrase`, which alleviates reliance on the `triplet` function and its limitations.
 * Added `lyrics` class that parallels the structure and behavior of the other classes; added associated functions and generic method implementations.
 * Added `lyrics` support to `music` object construction and transformations.
+* Added `render_*` functions for making simple sheet music snippets directly from `music` objects. This abstracts the `phrase() %>% track() [%>% trackbind()] %>% score() %>% render_*()` pipeline from the user for simpler music that is essentially a single voice, single track.
+* Added support for auto-cropping of rendered sheet music when the output format is png.
+* Added transparent background png support.
 * Added MIDI file read support (requires optional `tuneR` installation) and a set of functions for inspecting and manipulating the table of MIDI music data.
 * Added initial support for conversion of MIDI file input to `noteworthy`, `noteinfo`, `music` and `phrase` classes so the MIDI data can be analyzed, transformed, edited and rendered to sheet music and a new MIDI file.
 * Added more functions for music data manipulation and analysis.